# docs/openapi/operational-v1.yaml
openapi: 3.1.0
info:
  title: GiroMesa Operational API
  version: 2026-07-31
  description: Contracts introduced by operational redesign phase 1. Tenant is resolved by authenticated backend context.
paths:
...
  /api/v1/floor/reservations:
    post:
      summary: Create reservation optionally assigned to multiple tables
      responses:
        '201': { description: Reservation with tableIds and optimistic version }
        '409': { description: Table schedule conflict }
  /api/v1/floor/reservations/{reservationId}/seat:
    post:
      summary: Atomically seat reservation on assigned or supplied tables and open its order
      parameters:
        - in: path
          name: reservationId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201': { description: Seated reservation and active order }
        '409': { description: Concurrent occupancy conflict }
...